+++ core/ISliceBuffer.h
@@ -12,10 +12,10 @@ class ISliceBuffer {
 public:
     virtual void initialize() = 0;
     
-    virtual float read(int channel, long frame) = 0;
+    virtual float read(int channel, uint32_t frame) = 0;
     
     virtual void write(float in0, float in1) = 0;
-    virtual unsigned long writeHead() = 0;
+    virtual uint32_t writeHead() = 0;
     virtual void rewind() = 0;
     virtual bool isFull() = 0;
