+++ core/SliceBuffer.h
@@ -16,7 +16,7 @@ public:
     SliceBuffer();
 
     void initialize() override;
-    float read(int, long) override;
+    float read(int, uint32_t) override;
     void write(float, float) override;
     unsigned long writeHead() override { return _writeHead; }
     void rewind() override;
@@ -26,8 +26,8 @@ public:
     ~SliceBuffer() {};
     
 private:
-    unsigned long _size;
-    unsigned long _writeHead;
+    uint32_t _size;
+    uint32_t _writeHead;
     float* _buffer[2];
 };
